hello

+++ ft_printf.h
@@ -1,20 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_printf.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yoibarki <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/11/19 15:18:07 by yoibarki          #+#    #+#             */
+/*   Updated: 2022/11/19 15:18:11 by yoibarki         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
-#include <unistd.h>
-#include <stdarg.h>
-#include <stdlib.h>
+# include <stdarg.h>
+# include <stdio.h>
+# include <stdlib.h>
+# include <unistd.h>
+
 char	*ft_itoa(int n);
-int ft_printf(const char *s, ...);
-//static int checkformat(char c, va_list *varin);
-int ftt_count(int p);
-int ft_printp(int p);
-void ft_upplow(unsigned int h, char type);
-int ft_printhex(unsigned int h, char type);
-int ft_printunsigned(unsigned int nb);
-int ft_putchar(int a);
-int ft_prints(char *str);
-int ft_printnb(int nb);
-int ft_count(unsigned int h);
+int		ft_printf(const char *s, ...);
+int		ftt_count(unsigned long long p);
+int		ft_printp(unsigned long long p);
+void	ft_upplow(unsigned int h, char type);
+int		ft_printhex(unsigned int h, char type);
+int		ft_printunsigned(unsigned int nb);
+int		ft_putchar(int a);
+int		ft_prints(char *str);
+int		ft_printnb(int nb);
+int		ft_count(unsigned int h);
 
 #endif
